Support timeouts for routed requests when adding a method or a state.
Deamon implementations have to implement a timeout for routed request
(set/call). Otherwise, if a method/state owner does not respond to a routed
request, the routed request itself might linger forever in the jet
daemon. So every daemon has to implement a default timeout for routed
requests.

This extension allows a method/state owner to request a different timeout
value for this specific method/state to be added.

Daemons might ignore this timeout parameter.