Support timeouts for routed requests when setting a state or calling a method.
Deamon implementations have to implement a timeout for routed request
(set/call). Otherwise, if a state owner does not respond to a routed
request, the routed request itself might linger forever in the jet
daemon. So every daemon has to implement a default timeout for routed
requests.

This extension allows a peer to request a different timeout
value for this specific set() or call() method.

Jet daemons might ignore this requested timeout.